+++ src/clothes_shop/views/category_views.py
@@ -1,0 +1,89 @@
+import logging
+
+from django.shortcuts import get_object_or_404
+from rest_framework import generics, status
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+from clothes_shop.models import Brand, ClothesType, Size, Target
+from clothes_shop.serializers import (
+    BrandSerializer,
+    ClothesTypeSerializer,
+    SizeSerializer,
+    TargetSerializer,
+)
+
+logger = logging.getLogger(__name__)
+
+
+class SizeListCreateView(generics.ListCreateAPIView):
+    queryset = Size.objects.all()
+    serializer_class = SizeSerializer
+
+
+class SizeDetailView(generics.RetrieveUpdateDestroyAPIView):
+    queryset = Size.objects.all()
+    serializer_class = SizeSerializer
+
+    def get_object(self):
+        return get_object_or_404(Size, pk=self.kwargs.get("pk"))
+
+
+class TargetListCreateView(generics.ListCreateAPIView):
+    queryset = Target.objects.all()
+    serializer_class = TargetSerializer
+
+
+class TargetDetailView(generics.RetrieveUpdateDestroyAPIView):
+    queryset = Target.objects.all()
+    serializer_class = TargetSerializer
+
+    def get_object(self):
+        return get_object_or_404(Target, pk=self.kwargs.get("pk"))
+
+
+class ClothesTypeListCreateView(generics.ListCreateAPIView):
+    queryset = ClothesType.objects.all()
+    serializer_class = ClothesTypeSerializer
+
+
+class ClothesTypeDetailView(generics.RetrieveUpdateDestroyAPIView):
+    queryset = ClothesType.objects.all()
+    serializer_class = ClothesTypeSerializer
+
+    def get_object(self):
+        return get_object_or_404(ClothesType, pk=self.kwargs.get("pk"))
+
+
+class BrandListCreateView(generics.ListCreateAPIView):
+    queryset = Brand.objects.all()
+    serializer_class = BrandSerializer
+
+
+class BrandDetailView(generics.RetrieveUpdateDestroyAPIView):
+    queryset = Brand.objects.all()
+    serializer_class = BrandSerializer
+
+    def get_object(self):
+        return get_object_or_404(Brand, pk=self.kwargs.get("pk"))
+
+
+class CategoryListView(APIView):
+    def get(self, request):
+        sizes = Size.objects.all()
+        targets = Target.objects.all()
+        clothes_types = ClothesType.objects.all()
+        brands = Brand.objects.all()
+
+        size_serializer = SizeSerializer(sizes, many=True)
+        target_serializer = TargetSerializer(targets, many=True)
+        clothes_type_serializer = ClothesTypeSerializer(clothes_types, many=True)
+        brand_serializer = BrandSerializer(brands, many=True)
+
+        data = {
+            "sizeCatgory": size_serializer.data,
+            "targetCatgory": target_serializer.data,
+            "typeCatgory": clothes_type_serializer.data,
+            "brandCatgory": brand_serializer.data,
+        }
+        return Response(data, status=status.HTTP_200_OK)
